# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users/admin/stores/approve": {
      "post": {
        "operationId": "AdminController_aproveStore",
        "summary": "Approve a store",
        "description": "[Admin] Approve a store.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AproveRejectStroreDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The store has been successfully approved."
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "404": {
            "description": "The user is not found."
          },
          "422": {
            "description": "The store is already approved."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/admin/stores/reject": {
      "post": {
        "operationId": "AdminController_rejectStore",
        "summary": "Reject a store",
        "description": "[Admin] Reject a store.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AproveRejectStroreDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The store has been successfully rejected."
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "404": {
            "description": "The user is not found."
          },
          "422": {
            "description": "The store is already rejected."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/admin": {
      "post": {
        "operationId": "AdminController_create",
        "summary": "Create a new user",
        "description": "[Admin] Create a new user (default: admin).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The record has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "AdminController_findMany",
        "summary": "Get all the users by filtering them",
        "description": "[Admin] Get all the filtered users (doesn't return user type specific fields).",
        "parameters": [
          {
            "name": "user-type",
            "required": false,
            "in": "query",
            "description": "User Type (1: Admin, 2: Store, 3: Customer)",
            "schema": {
              "enum": [
                1,
                2,
                3
              ],
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Role (1: Admin, 2: Store, 3: Customer, 4: Unapproved Store, 5: Unconfirmed, 6: Guest)",
            "schema": {
              "enum": [
                1,
                2,
                3,
                4,
                5,
                6
              ],
              "type": "number"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NIF",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email-confirmed",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "The records have been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/admin/pagination": {
      "get": {
        "operationId": "AdminController_findManyWithPagination",
        "summary": "Get all the users by filtering them (with pagination)",
        "description": "[Admin] Get all the filtered users (with pagination) (doesn't return user type specific fields).",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Results/page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "user-type",
            "required": false,
            "in": "query",
            "description": "User Type (1: Admin, 2: Store, 3: Customer)",
            "schema": {
              "enum": [
                1,
                2,
                3
              ],
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Role (1: Admin, 2: Store, 3: Customer, 4: Unapproved Store, 5: Unconfirmed, 6: Guest)",
            "schema": {
              "enum": [
                1,
                2,
                3,
                4,
                5,
                6
              ],
              "type": "number"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "NIF",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email-confirmed",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "The records have been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/admin/{id}": {
      "get": {
        "operationId": "AdminController_findOne",
        "summary": "Get a user by ID",
        "description": "[Admin] Get the specified user by its ID (user_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "AdminController_update",
        "summary": "Update a user.",
        "description": "[Admin] Update the specified user by its ID (user_id) with the specified data.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The record has been successfully updated."
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_remove",
        "summary": "Delete a user",
        "description": "[Admin] Delete the specified user by its ID (user_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "description": "Delete mode (soft or hard). Default: soft",
            "schema": {
              "enum": [
                "soft",
                "hard"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully deleted."
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/admin/{id}/restore": {
      "patch": {
        "operationId": "AdminController_restore",
        "summary": "Restore a user",
        "description": "[Admin] Restore a soft deleted user specified  by its ID (user_id).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully restored."
          },
          "401": {
            "description": "The user is not logged in."
          },
          "403": {
            "description": "The user hasn't the right role (admin)."
          },
          "422": {
            "description": "The email is not valid."
          }
        },
        "tags": [
          "Users/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/books/admin/{ref}/restore": {
      "patch": {
        "operationId": "BooksAdminController_restoreOne",
        "summary": "Restore a book",
        "description": "[Admin] Restore a book from the store by its reference",
        "parameters": [
          {
            "name": "store_id",
            "required": false,
            "in": "query",
            "description": "ID of the store. Default 1",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "description": "Book reference",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The book has been successfully restored.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Book"
                    },
                    {
                      "properties": {
                        "languages": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Language"
                          }
                        }
                      }
                    },
                    {
                      "properties": {
                        "keywords": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Keyword"
                          }
                        }
                      }
                    },
                    {
                      "properties": {
                        "topics": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Topic"
                          }
                        }
                      }
                    },
                    {
                      "properties": {
                        "images": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Image"
                          }
                        }
                      }
                    },
                    {
                      "properties": {
                        "location": {
                          "$ref": "#/components/schemas/Location"
                        }
                      }
                    },
                    {
                      "properties": {
                        "status": {
                          "$ref": "#/components/schemas/Status"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Store not found."
          },
          "500": {
            "description": "The book has not been restored."
          }
        },
        "tags": [
          "Books/Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Librería Santa Bárbara API",
    "description": "API documentation for Librería Santa Bárbara",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AproveRejectStroreDto": {
        "type": "object",
        "properties": {
          "hash": {
            "type": "string",
            "example": "1a2b3c4d5e6f",
            "description": "The hash used to aprove or reject the store by the admin"
          }
        },
        "required": [
          "hash"
        ]
      },
      "CreateAdminDto": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "number",
            "example": 1,
            "description": "The role ID.",
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6
            ],
            "default": 1
          },
          "password": {
            "type": "string",
            "example": "test",
            "description": "The password of the store.",
            "minLength": 8
          },
          "address": {
            "type": "string",
            "example": "Calle Princesa 12B",
            "description": "The address of the user",
            "maxLength": 300
          },
          "phone_number": {
            "type": "string",
            "example": "+34661122334",
            "description": "The phone number of the user",
            "maxLength": 15
          },
          "email_confirmed": {
            "type": "boolean",
            "example": false,
            "default": false,
            "description": "Whether the email of the user has been confirmed or not by the himself."
          },
          "hash": {
            "type": "string",
            "example": null,
            "default": null,
            "description": "Hash used to verify the email of the user at registration"
          },
          "user_type_id": {
            "type": "number",
            "example": 1,
            "default": 1,
            "description": "The user type ID.",
            "enum": [
              1,
              2,
              3
            ]
          },
          "email": {
            "type": "string",
            "example": "admin@admin.com",
            "description": "The email of the admin. Should be unique.",
            "maxLength": 300
          },
          "NIF": {
            "type": "string",
            "example": null,
            "default": null,
            "description": "The NIF of the store. Should be alphanumeric and unique among stores.",
            "maxLength": 15
          }
        },
        "required": [
          "role_id",
          "password",
          "email"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the user"
          },
          "user_type_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the user type"
          },
          "role_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the role"
          },
          "email": {
            "type": "string",
            "example": "example@example.com",
            "description": "The email of the user"
          },
          "NIF": {
            "type": "string",
            "example": "12345678A",
            "description": "The NIF of the user"
          },
          "address": {
            "type": "string",
            "example": "Calle Princesa 12B",
            "description": "The address of the user"
          },
          "phone_number": {
            "type": "string",
            "example": "+34661122334",
            "description": "The phone number"
          },
          "email_confirmed": {
            "type": "boolean",
            "example": false,
            "description": "The email confirmation status of the user"
          },
          "hash": {
            "type": "string",
            "example": null,
            "description": "Hash used to verify email at registration"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2021-01-01",
            "description": "The date when the user was created"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "The date when the user has been updated"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string",
            "example": null,
            "description": "The date when the user has been (soft) deleted"
          }
        },
        "required": [
          "id",
          "user_type_id",
          "role_id",
          "email",
          "NIF",
          "address",
          "phone_number",
          "email_confirmed",
          "hash",
          "created_at",
          "updated_at",
          "deleted_at"
        ]
      },
...
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the language"
          },
          "book_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the book that the language references to"
          },
          "language": {
            "type": "string",
            "example": "EN",
            "description": "ISO 639-1 Alpha-2 code of the language"
          }
        },
        "required": [
          "id",
          "book_id",
          "language"
        ]
      },
...
      "Keyword": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the keyword"
          },
          "book_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the book that the keyword references to"
          },
          "keyword": {
            "type": "string",
            "example": "keyword",
            "description": "The keyword"
          }
        },
        "required": [
          "id",
          "book_id",
          "keyword"
        ]
      },
...
    }
  }
}